`ProfileBioPanelBuilder` rework complete. For now.
* Added `::build_native` class method in `Builder`, replacing `#to_ox` instance method;
* Broke `ProfileBioPanelBuilder`'s `Builder` class out to separate file;
* `ProfileBioPanelBuilder::Builder.build_native` has cleaner use of `Builder` methods;
* Removed no-longer-used view helper from parameters to `ProfileBioPanelBuilder`; ripples yielded cleaner code.

RSpec: 1062 examples, 0 failures, 1 pending
RuboCop: 302 files inspected, no offenses detected
Coverage: 3244 / 3257 LOC (99.6%) covered.
