+++ app/helpers/users_helper/profile_bio_row_builder.rb
@@ -16,23 +16,12 @@ class ProfileBioRowBuilder
     self
   end
 
-  # NOTE: PAY ATTENTION! WTF WORKAROUND IN PROGRESS!
-  # This started out as a fairly conventional Rails view-helper method, with the
-  # outer block below appearing as:
-  #     h.content_tag :div, nil, { class: 'row' }, false do
-  #       # innards pretty much as is
-  #     end
-  # The WTF is that, despite the very explicit `false` parameter to #content_tag
-  # (which should disable escaping), the content added by the block WAS escaped
-  # (and thus invalid as HTML).
-  #
-  # Nokogiri, the Swiss Army Nuclear Ginsu Chainsaw, to the rescue!
   Contract None => String
   def to_html
     outer_div = outer_container do |div|
       div['class'] = 'row'
       div << ProfileBioHeaderBuilder.new(user_name, h).to_html
-      div << ProfileBioPanelBuilder.new(user_profile, h).to_html
+      div << ProfileBioPanelBuilder.new(user_profile).to_html
     end
     reformat_html_output outer_div
   end
